refactor: flatten components
